adding movies and watchlist to crave
