+++ package.json
@@ -2,7 +2,10 @@
   "name": "react",
   "version": "1.0.0",
   "description": "React example starter project",
-  "keywords": ["react", "starter"],
+  "keywords": [
+    "react",
+    "starter"
+  ],
   "main": "src/index.js",
   "dependencies": {
     "react": "18.0.0",
@@ -11,6 +14,7 @@
   },
   "devDependencies": {
     "@babel/runtime": "7.13.8",
+    "react-app-alias": "^2.2.2",
     "typescript": "4.1.3"
   },
   "scripts": {
@@ -19,5 +23,10 @@
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
   },
-  "browserslist": [">0.2%", "not dead", "not ie <= 11", "not op_mini all"]
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
