+++ src/api/mock.js
@@ -1,0 +1,182 @@
+export const mock = [
+  {
+    id: 52061,
+    title: "Our Flag Means Death",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_2_25_25b23590-2669-48f5-8768-7a740d8c5744_jpg_2000x3000.jpg"
+  },
+  {
+    id: 50868,
+    title: "Shining Vale",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_2_14_59abaa4c-7af8-4243-bec0-839bb475a65c_png_2000x3000.jpg"
+  },
+  {
+    id: 45431,
+    title: "Friends",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_3_9_91c9ffc5-b3fa-4780-a776-de6989ca25ae_png_2000x3000.jpg"
+  },
+  {
+    id: 52101,
+    title: "Letterkenny: International Women's Day",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_3_8_0e28de35-baed-4928-b133-a4fcc77c6112_jpg_2000x3000.jpg"
+  },
+  {
+    id: 52028,
+    title: "Winning Time: The Rise of the Lakers Dynasty",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_3_2_6a9434e5-58bd-46c3-b3ab-1410cf401e40_jpg_2000x3000.jpg"
+  },
+  {
+    id: 50340,
+    title: "Hacks",
+    image:
+      "https://images2.9c9media.com/image_asset/2021_8_26_982cd5e5-a7fc-461f-8fe9-e7ba33986482_png_2000x3000.jpg"
+  },
+  {
+    id: 34522,
+    title: "South Park",
+    image:
+      "https://images2.9c9media.com/image_asset/2019_6_25_0912d848-4ae0-482f-8841-0b24b1b1220c_png_2000x3000.jpg"
+  },
+  {
+    id: 46364,
+    title: "Dollface",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_2_8_539766e6-966d-45a0-8ef3-7487f0c096e6_png_2000x3000.jpg"
+  },
+  {
+    id: 34584,
+    title: "Letterkenny",
+    image:
+      "https://images2.9c9media.com/image_asset/2021_12_17_64f49892-adc2-4ccb-b19d-495eb550098d_png_2000x3000.jpg"
+  },
+  {
+    id: 51732,
+    title: "Peacemaker",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_1_10_d5dfcca2-be82-40a5-93e3-dd25c94ba118_png_2000x3000.jpg"
+  },
+  {
+    id: 37701,
+    title: "Last Week Tonight With John Oliver",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_2_16_31dddb6c-9c6f-4516-8556-97af6527cc8d_png_2000x3000.jpg"
+  },
+  {
+    id: 45674,
+    title: "RuPaul's Drag Race",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_1_4_6761c172-9654-4de5-bd55-103abe01bcb9_png_2000x3000.jpg"
+  },
+  {
+    id: 45589,
+    title: "The Righteous Gemstones",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_1_6_0ebdcdc0-89db-4bb0-99e6-6db2b793f78b_png_2000x3000.jpg"
+  },
+  {
+    id: 50337,
+    title: "Children Ruin Everything",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_2_21_3e517210-17ab-4f2a-959c-47d8259ac07f_png_2000x3000.jpg"
+  },
+  {
+    id: 51840,
+    title: "Ride the Eagle",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_2_11_5a588b08-ea79-468d-851c-3caf422e8f03_png_2000x3000.jpg"
+  },
+  {
+    id: 51727,
+    title: "Somebody Somewhere",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_1_10_93be06fb-35d9-4264-8641-c82150b19c4e_png_2000x3000.jpg"
+  },
+  {
+    id: 48356,
+    title: "The Fresh Prince of Bel-Air",
+    image:
+      "https://images2.9c9media.com/image_asset/2020_11_9_e69d1277-e0ba-4786-92b2-f2184f3a6975_png_2000x3000.jpg"
+  },
+  {
+    id: 51718,
+    title: "Astrid and Lilly Save the World",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_1_24_af845b5f-f280-4145-a9eb-99a208d50cec_png_2000x3000.jpg"
+  },
+  {
+    id: 50430,
+    title: "The White Lotus",
+    image:
+      "https://images2.9c9media.com/image_asset/2021_8_26_643d2668-36d8-4f19-8b06-22a92904cfc1_png_2000x3000.jpg"
+  },
+  {
+    id: 50201,
+    title: "Wipeout",
+    image:
+      "https://images2.9c9media.com/image_asset/2021_5_12_39a00768-ae3d-4072-90fa-57a32c91574c_png_2000x3000.jpg"
+  },
+  {
+    id: 50410,
+    title: "Friends: The Reunion",
+    image:
+      "https://images2.9c9media.com/image_asset/2021_12_8_22ed44f3-d115-4147-9a09-ce75933a0a5a_png_2000x3000.jpg"
+  },
+  {
+    id: 29964,
+    title: "The Big Bang Theory",
+    image:
+      "https://images2.9c9media.com/image_asset/2021_2_10_e51a1e22-cce2-4882-9a01-39c248323671_png_2000x3000.jpg"
+  },
+  {
+    id: 51898,
+    title: "Pillow Talk",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_2_7_63cfd872-e2c6-4e05-8ec4-2277d76d00df_png_2000x3000.jpg"
+  },
+  {
+    id: 50702,
+    title: "Free Guy",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_2_2_6f62354b-1656-44f8-8713-43e588615890_png_2000x3000.jpg"
+  },
+  {
+    id: 45467,
+    title: "A Black Lady Sketch Show",
+    image:
+      "https://images2.9c9media.com/image_asset/2021_8_26_9eb5c57f-d21a-4281-b830-6409a590de84_png_2000x3000.jpg"
+  },
+  {
+    id: 51792,
+    title: "Dream Horse",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_2_11_ec4a34d9-98fb-4eca-bd4e-20e0ebfaa8b6_png_2000x3000.jpg"
+  },
+  {
+    id: 36834,
+    title: "Bridget Everett: Gynecological Wonder",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_2_4_31e891c5-e014-4898-ac41-8be029044360_png_2000x3000.jpg"
+  },
+  {
+    id: 51960,
+    title: "RuPaul's Drag Race: UK vs The World",
+    image:
+      "https://images2.9c9media.com/image_asset/2022_1_31_fe6f0609-8f89-4811-917b-488584f7daf9_png_2000x3000.jpg"
+  },
+  {
+    id: 48173,
+    title: "The Bernie Mac Show",
+    image:
+      "https://images2.9c9media.com/image_asset/2020_10_27_4d1bcf2f-5aa8-4900-aab0-81695c7e0642_png_2000x3000.jpg"
+  },
+  {
+    id: 37718,
+    title: "Real Time With Bill Maher",
+    image:
+      "https://images2.9c9media.com/image_asset/2021_8_26_2c247d30-2a92-4bea-9d31-6ddb5c628063_png_2000x3000.jpg"
+  }
+];
