+++ src/styles.css
@@ -2,11 +2,60 @@
   margin: 0;
   padding: 0fr;
   background-color: #212529;
+  color: white;
 }
+
 main {
   font-family: sans-serif;
 }
+
 header {
   background-color: black;
   padding: 16px 4vw;
+}
+
+.movie-cards {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
+  grid-gap: 60px 15px;
+  padding: 50px;
+  margin: 30px 0 30px 0;
+}
+
+.movie-image {
+  width: 100%;
+  height: auto;
+}
+
+.title {
+  margin: 7px auto 7px auto;
+  font-size: 16px;
+  height: 36px;
+  font-weight:bold;
+}
+
+.add-remove-btn {
+  font-size: 16px;
+  border: none;
+}
+
+.add-remove-btn:hover {
+  cursor: pointer;
+  color: rgb(0, 193, 243);
+}
+
+.loading {
+  position: fixed;
+  top: 50%;
+  left:50%;
+  transform: translate(-50%, -50%);
+}
+
+.movie-card {
+  padding: 10px;
+  transition: 0.3s ease-in-out;
+}
+
+.movie-card:hover {
+  transform: scale(1.04);
 }
